+++ lib/doorkeeper/openid_connect/user_info.rb
@@ -13,7 +13,7 @@ module Doorkeeper
       end
 
       def as_json(*_)
-        claims
+        claims.reject { |_, value| value.nil? || value == '' }
       end
 
       private